Splitting DDS structures into contrib file
- From BTDX archive
