+++ bethesda_structs/contrib/__init__.py
@@ -1,0 +1,2 @@
+# Copyright (c) 2018 Stephen Bunn <stephen@bunn.io>
+# GPLv3 License <https://choosealicense.com/licenses/gpl-3.0/>
